+++ kinematics/jacobianInverseKinematics_test.go
@@ -22,7 +22,7 @@ func TestCreateJacIKSolver(t *testing.T) {
 	pos.Theta *= math.Pi / 180
 
 	transform := kinmath.NewQuatTransFromRotation(pos.Theta, pos.RX, pos.RY, pos.RZ)
-	transform.SetTranslation(float64(pos.X), float64(pos.Y), float64(pos.Z))
+	transform.SetTranslation(pos.X, pos.Y, pos.Z)
 
 	ik.AddGoal(transform, 0)
 	solved := ik.Solve()
